Move the code for checking CR deletion into Reconcile
This commit moves the code for handling CR deletion from Process() to Reconcile(). If a CR is deleted while it is being processed, the Reconcile() method will be called and the CR is removed from the admissionMap. This will enable the asynchronously running Process() to check whether the CR is deleted periodically.

Testing Done:

- Unit tests pass.
- Created a CR. While it was processing, delete it. Ensured that the code for handling the delete succeeded.