Prevent concurrent CRs for one ASG
This commit introduces the inProcessASGs map for keeping track of which ASGs are being processed. If an ASG is being processed and a new CR is submitted for the same ASG, the new CR get requeued until the first one is completed.

This takes care of the case when a CR is deleted in the middle of doing rolling-upgrade. If the CR is deleted, no new instances will get drained and terminated. The ones in the middle of drain and terminate will be completed.

Testing Done:

- Adjusted unit tests for the new code. Ensured that unit tests complete successfully.
- Deleted rollup CR when it was in the middle of being processed. Ensured that the CR completed the node that was being processed. Others nodes were ignored correctly.
- Deleted rollup CR after it was completed. Ensured that it was deleted.