Check for CR before updating each instance. (#97)
* Mark an instance as "completed" iff it was "in-progress"

If an instance was started to get updated, it was marked "in-progress". However, there are cases when the instance is never marked "completed". This commit fixes that.

All instance are marked "completed" if they started their upgrade.

Testing Done:

- Unit tests for this exist. Modified them to confirm expected behavior.

* Add rolling upgrade CR name to admissionMap before processing starts

This ensures that the admissionMap could be used in subsequent go-routines. In later commits, the admissionMap will be updated to indicate that the CR has been deleted. The go-routines that update the instances will check this and choose to proceed or not.

Testing Done:

- Unit tests pass. Ran a simple upgrade. It succeeded.

* Remove return values from "Process()"

This is in preparation for calling Process() in a go-routine.

Testing Done:
- Unit tests pass.

* Make Process run asynchronously

This enables the Reconcile() to be called again. If and when the CR gets modified/deleted, the Reconcile() can now be called and appropriate actions can be taken.

Testing Done:

- Unit tests pass
- Tested basic rolling-upgrade.

* Move the code for checking CR deletion into Reconcile

This commit moves the code for handling CR deletion from Process() to Reconcile(). If a CR is deleted while it is being processed, the Reconcile() method will be called and the CR is removed from the admissionMap. This will enable the asynchronously running Process() to check whether the CR is deleted periodically.

Testing Done:

- Unit tests pass.
- Created a CR. While it was processing, delete it. Ensured that the code for handling the delete succeeded.

* Check admissionMap for every UpdateInstance

This commit adds a check for an entry in the admissionMap during every UpdateInstance run. If the CR is deleted, the entry for it is removed from the admissionMap. And any calls to UpdateInstance from then on get short-circuited.

Testing Done:
- Unit tests passed
- Basic rolling upgrade worked as expected.

* Ignore "Invalid object" errors during status update.

These errors are thrown if the CR is deleted while it was being processed. Unfortunately, there is no way other than error string comparison for doing this.

Testing Done:

- Unit tests succeed.
- Started an upgrade. Deleted the CR while being processed. Verified that the correct log messages were logged and the controller didn't crash.

* Prevent concurrent CRs for one ASG

This commit introduces the inProcessASGs map for keeping track of which ASGs are being processed. If an ASG is being processed and a new CR is submitted for the same ASG, the new CR get requeued until the first one is completed.

This takes care of the case when a CR is deleted in the middle of doing rolling-upgrade. If the CR is deleted, no new instances will get drained and terminated. The ones in the middle of drain and terminate will be completed.

Testing Done:

- Adjusted unit tests for the new code. Ensured that unit tests complete successfully.
- Deleted rollup CR when it was in the middle of being processed. Ensured that the CR completed the node that was being processed. Others nodes were ignored correctly.
- Deleted rollup CR after it was completed. Ensured that it was deleted.

* Best effort instance tagging after termination.