Fix log arguments for 0 node ASGs
This commit fixes the arguments sent to the logger when logging a message about 0 nodes in the ASG. Prior to this commit, there were odd number of arguments in the call and caused it to crash. After the change, the log message is logged correctly.

Unrelated but required change:
- Increased the size of the swapfile for resolving OutOfMemory build errors.

Testing Done:

- Reproduced the problem reported in #104. Found the controller crashing. Changed the docker image to a new one which included this fix.
- Verified that the fix worked as expected.

Closes #104